requirements.txt
ago==0.0.95
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
boto3==1.28.75
botocore==1.31.75
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
dateparser==1.1.8
dotmap==1.3.30
elastic-transport==8.10.0
elasticsearch==8.10.1
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.13.1
hjson==3.1.0
hurry.filesize==0.9
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.3.2
langdetect==1.0.9
lxml==4.9.3
news-please==1.5.35
newspaper3k==0.2.8
nltk==3.8.1
packaging==23.2
parsel==1.8.1
Pillow==10.1.0
plac==1.4.1
plotille==5.0.0
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
s3transfer==0.7.0
Scrapy==2.11.0
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
stanza==1.8.1
tinysegmenter==0.3
tldextract==5.0.1
tqdm==4.66.1
Twisted==22.10.0
#typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
w3lib==2.1.2
warcio==1.7.4
zope.interface==6.1
sentence-transformers
wordfreq
h5py
tables
pykdtree==1.3.5
markov-clustering